tests: Honor GBP_TESTS_NOCLEAN in rpm tests too

+++ tests/20_test_rpm.py
@@ -56,7 +56,8 @@ class RpmTestBase(object):
 
     def teardown(self):
         """Test case teardown"""
-        shutil.rmtree(self.tmpdir)
+        if not os.getenv("GBP_TESTS_NOCLEAN"):
+            shutil.rmtree(self.tmpdir)
 
 
 class TestSrcRpmFile(RpmTestBase):
